// src/base/theme.d.ts
declare const theme: {
  brand: {
    main: string;
    dark: string;
    light: string;
  };
  commons: {
    dark: string;
    main: string;
    light: string;
  };
  fontFamily: string;
  default: {
    color: {
      main: string;
    };
    border: {
      main: string;
    };
    background: {
      main: string;
      hover: string;
      focus: string;
    };
  };
  dark: {
    color: {
      main: string;
      hover: string;
      focus: string;
    };
    border: {
      main: string;
      hover: string;
      focus: string;
    };
    background: {
      main: string;
    };
  };
  light: {
    color: {
      main: string;
      hover: string;
      focus: string;
    };
    border: {
      main: string;
      hover: string;
      focus: string;
    };
    background: {
      main: string;
    };
  };
};
export default theme;